be/requirements/base.txt
wheel==0.40.0
asgiref==3.7.2
django==4.2.3
sqlparse==0.4.2
djangorestframework==3.14.0
django-environ==0.10.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
mysqlclient==2.2.0
langchain==0.0.244
openai==0.27.8
tiktoken==0.4.0
chromadb==0.4.2
pypdf==3.12.2
pymupdf==1.23.5
azure-storage-blob==12.17.0
django-storages[azure]==1.13.2
typing-inspect==0.8.0
typing_extensions==4.5.0
pydantic==1.10.8
pysqlite3-binary==0.5.0
unstructured==0.8.4
azure-search-documents==11.4.0b6
azure-identity==1.13.0